Create README.md
This program extracts grades from a list in the format ('studentNum-Grade%') by using the split and strip methods. The program then converts the extracted string into a float so that it can be used in further calculations. From the extracted grades, the program is then able to calculate the class average as well as the standard deviation grade.